requirements.txt
boto3>=1.26.19
requests>=2.28.1
mysqlclient>=2.1.1